refactor(ai-chat): Split ai chat screen into per widget files
