+++ trakli-mobile.code-workspace
@@ -1,0 +1,14 @@
+{
+	"folders": [
+		{
+			"path": "."
+		},
+		{
+			"path": "../webservice"
+		},
+		{
+			"path": "../webui"
+		}
+	],
+	"settings": {}
+}
